requirements.txt
amqp==5.1.1
asgiref==3.5.2
Babel==2.10.1
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==4.0.4
django-bootstrap-modal-forms==2.2.0
django-bootstrap3==21.2
django-celery-beat==2.3.0
django-crispy-forms==1.14.0
django-environ==0.8.1
django-filter==21.1
django-localflavor==3.1
django-money==2.1.1
django-tables2==2.4.1
django-timezone-field==5.0
environ==1.0
form==0.0.1
kombu==5.2.4
mysqlclient==2.1.0
prompt-toolkit==3.0.30
py-moneyed==1.2
python-crontab==2.6.0
python-dateutil==2.8.2
python-stdnum==1.17
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
vine==5.0.0
wcwidth==0.2.5
